+++ BE/src/main/java/com/codesquad/airbnb/common/exception/GlobalRestExceptionHandler.java
@@ -6,14 +6,21 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
 import java.nio.file.AccessDeniedException;
 
 @Slf4j
-@RestController
+@RestControllerAdvice
 public class GlobalRestExceptionHandler {
+
+    @ExceptionHandler(IllegalReservationDateException.class)
+    protected ResponseEntity<ErrorResponse> handleInputMistakeException(IllegalReservationDateException e) {
+        log.error("handleInputMistakeException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_INPUT_VALUE);
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
