+++ BE/src/main/java/com/codesquad/airbnb/common/exception/IllegalReservationDateException.java
@@ -1,0 +1,12 @@
+package com.codesquad.airbnb.common.exception;
+
+public class IllegalReservationDateException extends RuntimeException {
+
+    private final String name;
+
+    public static final String ERROR_CODE = "Input"
+
+    public IllegalReservationDateException(String message) {
+        super(message);
+    }
+}
