--- BE/src/main/java/com/codesquad/airbnb/common/exception/InputMistakeException.java
@@ -1,8 +1,0 @@
-package com.codesquad.airbnb.common.exception;
-
-public class InputMistakeException extends RuntimeException {
-
-    public InputMistakeException(String message) {
-        super(message);
-    }
-}
