--- BE/src/main/java/com/codesquad/airbnb/common/exception/InputNothingException.java
@@ -1,4 +1,0 @@
-package com.codesquad.airbnb.common.exception;
-
-public class InputNothingException {
-}
